preparing to compare the new approach (voxelization2) with current.
